apps/api/requirements.txt

fastapi==0.115.6
uvicorn==0.32.0
pydantic==2.9.2
PyJWT==2.9.0
pytest==8.3.3
